requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
factory_boy==3.3.1
Faker==35.0.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
